backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
alembic==1.13.3
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic-settings==2.5.2
email-validator==2.2.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis[hiredis]==5.0.8
elasticsearch[async]==8.15.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
asgi-lifespan==2.1.0
